+++ Page.h
@@ -5,7 +5,7 @@ namespace virtual_memory {
     class Page {
     public:
         Page() = default;
-        Page(int id, int number_of_page, double timestamp, int arrival_number, bool is_free);
+        Page(int id, int number_of_page, int timestamp, int arrival_number, bool is_free);
 
         int GetId() const ;
         int GetNumberOfPage() const;
@@ -16,7 +16,7 @@ namespace virtual_memory {
     private:
         int id_;
         int number_of_page_;
-        double timestamp_;
+        int timestamp_;
         int arrival_number_;
         bool is_free_;
     };
